move checkbox component to common folder
