+++ bigbluebutton-html5/imports/ui/components/common/checkbox/component.jsx
