+++ bigbluebutton-html5/imports/ui/components/presentation/presentation-uploader/component.jsx
@@ -13,7 +13,7 @@ import _ from 'lodash';
 import { registerTitleView, unregisterTitleView } from '/imports/utils/dom-utils';
 import Styled from './styles';
 import Settings from '/imports/ui/services/settings';
-import Checkbox from '/imports/ui/components/checkbox/component';
+import Checkbox from '/imports/ui/components/common/checkbox/component';
 
 const { isMobile } = deviceInfo;
